third_party/blink/web_tests/http/tests/security/host-compare-case-insensitive-expected.txt

This tests that the comparison of the host for security is done case insensitive.

 

--------
Frame: 'frameA'
--------
PASS: Host name compare is case insensitive.

Inner iframe.

--------
Frame: 'frameB'
--------
Other inner iframe.
